WIP educe.stac document-centric feature extraction
